Add conditional building to remaining storage APIs
