+++ federationsender/storage/interface.go
@@ -12,29 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build !wasm
-
 package storage
 
 import (
-	"net/url"
+	"context"
 
-	"github.com/matrix-org/dendrite/federationsender/storage/postgres"
-	"github.com/matrix-org/dendrite/federationsender/storage/sqlite3"
+	"github.com/matrix-org/dendrite/common"
+	"github.com/matrix-org/dendrite/federationsender/types"
 )
 
-// NewDatabase opens a new database
-func NewDatabase(dataSourceName string) (Database, error) {
-	uri, err := url.Parse(dataSourceName)
-	if err != nil {
-		return postgres.NewDatabase(dataSourceName)
-	}
-	switch uri.Scheme {
-	case "file":
-		return sqlite3.NewDatabase(dataSourceName)
-	case "postgres":
-		return postgres.NewDatabase(dataSourceName)
-	default:
-		return postgres.NewDatabase(dataSourceName)
-	}
+type Database interface {
+	common.PartitionStorer
+	UpdateRoom(ctx context.Context, roomID, oldEventID, newEventID string, addHosts []types.JoinedHost, removeHosts []string) (joinedHosts []types.JoinedHost, err error)
+	GetJoinedHosts(ctx context.Context, roomID string) ([]types.JoinedHost, error)
 }
