+++ federationsender/storage/storage_wasm.go
@@ -12,14 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build !wasm
-
 package storage
 
 import (
+	"fmt"
 	"net/url"
 
-	"github.com/matrix-org/dendrite/federationsender/storage/postgres"
 	"github.com/matrix-org/dendrite/federationsender/storage/sqlite3"
 )
 
@@ -27,14 +25,14 @@ import (
 func NewDatabase(dataSourceName string) (Database, error) {
 	uri, err := url.Parse(dataSourceName)
 	if err != nil {
-		return postgres.NewDatabase(dataSourceName)
+		return nil, fmt.Errorf("Cannot use postgres implementation")
 	}
 	switch uri.Scheme {
 	case "file":
 		return sqlite3.NewDatabase(dataSourceName)
 	case "postgres":
-		return postgres.NewDatabase(dataSourceName)
+		return nil, fmt.Errorf("Cannot use postgres implementation")
 	default:
-		return postgres.NewDatabase(dataSourceName)
+		return nil, fmt.Errorf("Cannot use postgres implementation")
 	}
 }
